+++ example/android/app/src/main/java/com/example/rnwidget/MainActivity.java
@@ -1,8 +1,14 @@
 package com.example.rnwidget;
 
+import android.os.Bundle;
 import com.facebook.react.ReactActivity;
 
 public class MainActivity extends ReactActivity {
+
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    super.onCreate(null);
+  }
 
   /**
    * Returns the name of the main component registered from JavaScript. This is used to schedule
