+++ example/package.json
@@ -9,10 +9,14 @@
     "start": "react-native start"
   },
   "dependencies": {
-    "react": "16.13.1",
-    "react-native": "0.63.4",
     "@react-native-community/datetimepicker": "^3.5.2",
+    "@react-navigation/native": "^6.0.10",
+    "@react-navigation/native-stack": "^6.6.2",
+    "react": "16.13.1",
+    "react-native": "0.64.3",
     "react-native-image-picker": "^4.3.0",
+    "react-native-safe-area-context": "^4.3.1",
+    "react-native-screens": "^3.13.1",
     "react-native-simple-toast": "^1.1.3"
   },
   "devDependencies": {
